test(backend): organize test files
