+++ backend/package.json
@@ -8,8 +8,7 @@
     "start": "nodemon --exec tsx server.ts",
     "scrape": "tsx scraper/scraper.ts",
     "scrape:ewant": "tsx scraper/scraperEwant.ts",
-    "test:course-time": "tsx --test utils/parseCourseTime.test.ts",
-    "test:timetable-analytics": "tsx --test models/GuestTimetableSnapshot.test.ts services/timetableAnalyticsService.test.ts repositories/timetableAnalyticsRepository.test.ts controllers/timetableAnalyticsController.test.ts routes/timetableAnalytics.test.ts",
+    "test": "tsx --test tests/**/*.test.ts",
     "backfill:schedules": "tsx scripts/backfillCourseSchedules.ts",
     "backfill:course-identity": "tsx scripts/backfillCourseIdentity.ts",
     "backfill:dcard-related-post-counts": "tsx scripts/backfillDcardRelatedPostCounts.ts",
