+++ backend/tests/timetableAnalytics/GuestTimetableSnapshot.model.test.ts
@@ -1,6 +1,6 @@
 import assert from "node:assert/strict";
 import test from "node:test";
-import GuestTimetableSnapshotModel from "./GuestTimetableSnapshot";
+import GuestTimetableSnapshotModel from "../../models/GuestTimetableSnapshot";
 
 test("GuestTimetableSnapshots model 使用單表 JSON snapshot 與必要索引", () => {
   assert.equal(GuestTimetableSnapshotModel.tableName, "GuestTimetableSnapshots");
